refactor: 도메인 및 테이블명 변경, 질문 서비스 메서드 리팩토링
- 제미나이 서비스를 제미나이 도메인으로 이동
- 사용하지 않는 도메인 및 파일 삭제
- 테이블명 변경
- 질문 서비스 메서드를 유스케이스별로 분리
- 메서드명 변경
