--- src/main/java/com/team/futureway/common/utils/QuestionKind.java
@@ -1,10 +1,0 @@
-package com.team.futureway.common.utils;
-
-import lombok.Getter;
-
-@Getter
-public enum QuestionKind {
-
-
-
-}
