+++ src/main/java/com/team/futureway/consult/entity/Question.java
@@ -1,4 +1,4 @@
-package com.team.futureway.gemini.entity;
+package com.team.futureway.consult.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -12,12 +12,12 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @Getter
 @Entity
-@Table(name = "ai_consultation_history")
-public class AiConsultationHistory {
+@Table(name = "question")
+public class Question {
     @Id
-    @Column(name = "ai_consultation_history_id")
+    @Column(name = "question_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long aiConsultationHistoryId;
+    private Long questionId;
 
     @Column(name = "user_id", nullable = false)
     private Long userId;
@@ -34,8 +34,8 @@ public class AiConsultationHistory {
     @Column(name = "created_date", nullable = false)
     private LocalDateTime createdDate; // 생성 일자
 
-    public static AiConsultationHistory of(Long aiConsultationHistoryId, Long userId, int questionNumber, String questionMessage, String answer) {
-        return new AiConsultationHistory(aiConsultationHistoryId, userId, questionNumber, questionMessage, answer, LocalDateTime.now());
+    public static Question of(Long aiConsultationHistoryId, Long userId, int questionNumber, String questionMessage, String answer) {
+        return new Question(aiConsultationHistoryId, userId, questionNumber, questionMessage, answer, LocalDateTime.now());
     }
 
     public void incrementQuestionNumber() {
