+++ src/main/java/com/team/futureway/consult/repository/QuestionRepository.java
@@ -1,0 +1,10 @@
+package com.team.futureway.consult.repository;
+
+import com.team.futureway.consult.entity.Question;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface QuestionRepository extends JpaRepository<Question, Long> {
+    List<Question> findByUserId(Long userId);
+}
