+++ src/main/java/com/team/futureway/consult/repository/SummaryRepository.java
@@ -1,0 +1,7 @@
+package com.team.futureway.consult.repository;
+
+import com.team.futureway.consult.entity.Summary;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SummaryRepository extends JpaRepository<Summary, Long> {
+}
