--- src/main/java/com/team/futureway/consult/request/ConsultRequest.java
@@ -1,4 +1,0 @@
-package com.team.futureway.consult.request;
-
-public record ConsultRequest(Long userId, String kind, String interest) {
-}
