+++ src/main/java/com/team/futureway/consult/response/AnswerRequest.java
@@ -1,4 +1,4 @@
-package com.team.futureway.gemini.response;
+package com.team.futureway.consult.response;
 
 public record AnswerRequest(Long aiConsultationHistoryId, Long userId, String answer) {
 }
