+++ src/main/java/com/team/futureway/consult/response/SummaryRequest.java
@@ -1,4 +1,4 @@
-package com.team.futureway.gemini.response;
+package com.team.futureway.consult.response;
 
 public record SummaryRequest(Long userId) {
 }
