--- src/main/java/com/team/futureway/gemini/repository/AiConsultationSummaryHistoryRepository.java
@@ -1,8 +1,0 @@
-package com.team.futureway.gemini.repository;
-
-import com.team.futureway.gemini.entity.AiConsultationSummaryHistory;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AiConsultationSummaryHistoryRepository extends JpaRepository<AiConsultationSummaryHistory, Long>{
-
-}
