+++ src/main/java/com/team/futureway/gemini/util/ParseGeminiMessageUtil.java
@@ -1,6 +1,5 @@
 package com.team.futureway.gemini.util;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.stereotype.Component;
