+++ src/main/java/com/team/futureway/user/controller/UserController.java
@@ -1,9 +1,13 @@
 package com.team.futureway.user.controller;
 
 import com.team.futureway.user.dto.UserDTO;
+import com.team.futureway.user.dto.UserTypeDTO;
 import com.team.futureway.user.response.UserRequest;
 import com.team.futureway.user.response.UserResponse;
+import com.team.futureway.user.response.UserTypeRequest;
+import com.team.futureway.user.response.UserTypeResponse;
 import com.team.futureway.user.service.UserService;
+import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -19,6 +23,20 @@ public class UserController {
     public ResponseEntity<UserResponse> save(@RequestBody UserRequest request) {
         UserDTO userDTO = userService.save(request.name());
         return ResponseEntity.ok(UserResponse.of(userDTO));
+    }
+
+    @Operation(summary = "사용자 유형 저장")
+    @PostMapping("/type")
+    public ResponseEntity<UserTypeResponse> type(@RequestBody UserTypeRequest request) {
+        UserTypeDTO userTypeDTO = UserTypeDTO.of(
+                request.userId(),
+                request.question(),
+                request.selectType(),
+                request.answer(),
+                request.userType()
+        );
+        UserTypeDTO result = userService.saveUserType(userTypeDTO);
+        return ResponseEntity.ok(UserTypeResponse.of(result));
     }
 
 }
