+++ src/main/java/com/team/futureway/user/response/UserTypeRequest.java
@@ -1,4 +1,4 @@
-package com.team.futureway.gemini.response;
+package com.team.futureway.user.response;
 
 public record UserTypeRequest(Long userId, String question, String selectType, String answer, String userType) {
 }
